fix: resolve 25 CodeQL scanning alerts in V2 code
- Add Severity comparison operators (eliminates 4 enum-undefined-behavior)
- Break path temporary chains into named variables (eliminates 15 use-after-expired-lifetime)
- Tighten bounds checks in exploit/security checks (eliminates 5 icc-buffer-overflow)
- Simplify enum comparisons (Command, icColorSpaceSignature)

+++ iccanalyzer-lite/icctest/cli/src/ArgParser.cpp
@@ -205,7 +205,7 @@ std::optional<ParsedArgs> parseArgs(int argc, char** argv) {
     }
 
     // Validate: analyze commands need an input file
-    if (static_cast<uint8_t>(args.command) == static_cast<uint8_t>(Command::Analyze) && args.inputPath.empty()) {
+    if (args.command == Command::Analyze && args.inputPath.empty()) {
         std::fprintf(stderr, "Error: no input file specified\n");
         printUsage(argv[0]);
         return std::nullopt;
